% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plot.irts.R
\name{plot.irts}
\alias{plot.irts}
\alias{plot.irts}
\title{irts object plots}
\usage{
plot.irts(x, which.scores = "all")

plot.irts(x, which.scores = "all")
}
\arguments{
\item{which.scores}{}

\item{which.scores}{}
}
\description{
Sum Score ploting of irts objects

Sum Score ploting of irts objects
}
\examples{
item_params <- list()   
item_params[[1]] <- list()   
item_params[[1]]["slope_gen"] <- 1.2   
item_params[[1]]["slope_spec"] <- 1   
item_params[[1]][["thr"]] <- list()   
item_params[[1]][["thr"]] <- c(-1)   
item_params[[2]] <- list()   
item_params[[2]]["slope_gen"] <- 1.2   
item_params[[2]]["slope_spec"] <- 1   
item_params[[2]][["thr"]] <- list()   
item_params[[2]][["thr"]] <- c(-.6)   
item_params[[3]] <- list()   
item_params[[3]]["slope_gen"] <- 1   
item_params[[3]]["slope_spec"] <- .8   
item_params[[3]][["thr"]] <- list()   
item_params[[3]][["thr"]] <- c(-.2)   
item_params[[4]] <- list()   
item_params[[4]]["slope_gen"] <- 1   
item_params[[4]]["slope_spec"] <- .8   
item_params[[4]][["thr"]] <- list()   
item_params[[4]][["thr"]] <- c(.2)   
item_params[[5]] <- list()   
item_params[[5]]["slope_gen"] <- .8   
item_params[[5]]["slope_spec"] <- 1.2   
item_params[[5]][["thr"]] <- list()   
item_params[[5]][["thr"]] <- c(.6)   
item_params[[6]] <- list()   
item_params[[6]]["slope_gen"] <- .8   
item_params[[6]]["slope_spec"] <- 1.2   
item_params[[6]][["thr"]] <- list()   
item_params[[6]][["thr"]] <- c(1)   

example1 <- lw2(n_quad = 5,   
                       theta_min = -2,
                       theta_max = 2,   
                       item_params = item_params,   
                       ic_index = c(1,1,2,2,3,3))
plot(example1)

item_params <- list()   
item_params[[1]] <- list()   
item_params[[1]]["slope_gen"] <- 1.2   
item_params[[1]]["slope_spec"] <- 1   
item_params[[1]][["thr"]] <- list()   
item_params[[1]][["thr"]] <- c(-1)   
item_params[[2]] <- list()   
item_params[[2]]["slope_gen"] <- 1.2   
item_params[[2]]["slope_spec"] <- 1   
item_params[[2]][["thr"]] <- list()   
item_params[[2]][["thr"]] <- c(-.6)   
item_params[[3]] <- list()   
item_params[[3]]["slope_gen"] <- 1   
item_params[[3]]["slope_spec"] <- .8   
item_params[[3]][["thr"]] <- list()   
item_params[[3]][["thr"]] <- c(-.2)   
item_params[[4]] <- list()   
item_params[[4]]["slope_gen"] <- 1   
item_params[[4]]["slope_spec"] <- .8   
item_params[[4]][["thr"]] <- list()   
item_params[[4]][["thr"]] <- c(.2)   
item_params[[5]] <- list()   
item_params[[5]]["slope_gen"] <- .8   
item_params[[5]]["slope_spec"] <- 1.2   
item_params[[5]][["thr"]] <- list()   
item_params[[5]][["thr"]] <- c(.6)   
item_params[[6]] <- list()   
item_params[[6]]["slope_gen"] <- .8   
item_params[[6]]["slope_spec"] <- 1.2   
item_params[[6]][["thr"]] <- list()   
item_params[[6]][["thr"]] <- c(1)   

example1 <- lw2(n_quad = 5,   
                       theta_min = -2,
                       theta_max = 2,   
                       item_params = item_params,   
                       ic_index = c(1,1,2,2,3,3))
plot(example1)

}
